vite/template/template/vite-starter-template-js/server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = ""
authors = [
    {name = "Umar Hayat",email = "umarhayat.coder@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard] (>=0.115.8,<0.116.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "pydantic (>=2.10.6,<3.0.0)",
    "passlib (>=1.7.4,<2.0.0)",
    "pyjwt (>=2.10.1,<3.0.0)",
    "authlib (>=1.4.1,<2.0.0)",
    "pyotp (>=2.9.0,<3.0.0)",
    "pymongo (>=4.11.1,<5.0.0)",
    "uvicorn (>=0.34.0,<0.35.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
